kvcoord: break dependency on localtestcluster
The earlier form prevented localtestcluster to transitively depend on
kvcoord, something we want in #73876 through the rangefeed package.
This commit moves a few tests to package external ones; to do so it
exports a few symbols (cluster setting variables) and defines some
(public) testing methods to mutate/retrieve internal state.

Release note: None
