+++ pkg/kv/kvclient/kvcoord/BUILD.bazel
@@ -129,6 +129,7 @@ go_test(
         "txn_coord_sender_server_test.go",
         "txn_coord_sender_test.go",
         "txn_correctness_test.go",
+        "txn_intercepter_pipeliner_client_test.go",
         "txn_interceptor_committer_test.go",
         "txn_interceptor_heartbeater_test.go",
         "txn_interceptor_pipeliner_test.go",
