+++ pkg/kv/kvclient/kvcoord/txn_interceptor_pipeliner_test.go
@@ -14,12 +14,10 @@ import (
 	"context"
 	"fmt"
 	"math"
-	"reflect"
 	"strings"
 	"testing"
 	"time"
 
-	"github.com/cockroachdb/cockroach/pkg/kv"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/concurrency/lock"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
@@ -585,7 +583,7 @@ func TestTxnPipelinerManyWrites(t *testing.T) {
 
 	// Disable write_pipelining_max_outstanding_size and max_intents_bytes limits.
 	pipelinedWritesMaxBatchSize.Override(ctx, &tp.st.SV, 0)
-	trackedWritesMaxSize.Override(ctx, &tp.st.SV, math.MaxInt64)
+	TrackedWritesMaxSize.Override(ctx, &tp.st.SV, math.MaxInt64)
 
 	const writes = 2048
 	keyBuf := roachpb.Key(strings.Repeat("a", writes+1))
@@ -986,7 +984,7 @@ func TestTxnPipelinerMaxInFlightSize(t *testing.T) {
 	tp, mockSender := makeMockTxnPipeliner(rangeIter)
 
 	// Set budget limit to 3 bytes.
-	trackedWritesMaxSize.Override(ctx, &tp.st.SV, 3)
+	TrackedWritesMaxSize.Override(ctx, &tp.st.SV, 3)
 
 	txn := makeTxnProto()
 	keyA, keyB := roachpb.Key("a"), roachpb.Key("b")
@@ -1129,7 +1127,7 @@ func TestTxnPipelinerMaxInFlightSize(t *testing.T) {
 	require.Equal(t, int64(0), tp.ifWrites.byteSize())
 
 	// Increase the budget limit to 5 bytes.
-	trackedWritesMaxSize.Override(ctx, &tp.st.SV, 5)
+	TrackedWritesMaxSize.Override(ctx, &tp.st.SV, 5)
 	tp.lockFootprint.clear() // hackily disregard the locks
 
 	// The original batch with 4 writes should succeed.
@@ -1495,144 +1493,6 @@ func TestTxnPipelinerSavepoints(t *testing.T) {
 	require.Empty(t, tp.ifWrites.len())
 }
 
-// TestTxnCoordSenderCondenseLockSpans verifies that lock spans are condensed
-// along range boundaries when they exceed the maximum intent bytes threshold.
-//
-// TODO(andrei): Merge this test into TestTxnPipelinerCondenseLockSpans2, which
-// uses a txnPipeliner instead of a full TxnCoordSender.
-func TestTxnPipelinerCondenseLockSpans(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	defer log.Scope(t).Close(t)
-	ctx := context.Background()
-	a := roachpb.Span{Key: roachpb.Key("a"), EndKey: roachpb.Key(nil)}
-	b := roachpb.Span{Key: roachpb.Key("b"), EndKey: roachpb.Key(nil)}
-	c := roachpb.Span{Key: roachpb.Key("c"), EndKey: roachpb.Key(nil)}
-	d := roachpb.Span{Key: roachpb.Key("ddddddd"), EndKey: roachpb.Key(nil)}
-	e := roachpb.Span{Key: roachpb.Key("e"), EndKey: roachpb.Key(nil)}
-	aToBClosed := roachpb.Span{Key: roachpb.Key("a"), EndKey: roachpb.Key("b").Next()}
-	cToEClosed := roachpb.Span{Key: roachpb.Key("c"), EndKey: roachpb.Key("e").Next()}
-	fTof0 := roachpb.Span{Key: roachpb.Key("f"), EndKey: roachpb.Key("f0")}
-	g := roachpb.Span{Key: roachpb.Key("g"), EndKey: roachpb.Key(nil)}
-	g0Tog1 := roachpb.Span{Key: roachpb.Key("g0"), EndKey: roachpb.Key("g1")}
-	fTog1Closed := roachpb.Span{Key: roachpb.Key("f"), EndKey: roachpb.Key("g1")}
-	testCases := []struct {
-		span         roachpb.Span
-		expLocks     []roachpb.Span
-		expLocksSize int64
-	}{
-		{span: a, expLocks: []roachpb.Span{a}, expLocksSize: 1},
-		{span: b, expLocks: []roachpb.Span{a, b}, expLocksSize: 2},
-		{span: c, expLocks: []roachpb.Span{a, b, c}, expLocksSize: 3},
-		{span: d, expLocks: []roachpb.Span{a, b, c, d}, expLocksSize: 10},
-		// Note that c-e condenses and then lists first.
-		{span: e, expLocks: []roachpb.Span{cToEClosed, a, b}, expLocksSize: 5},
-		{span: fTof0, expLocks: []roachpb.Span{cToEClosed, a, b, fTof0}, expLocksSize: 8},
-		{span: g, expLocks: []roachpb.Span{cToEClosed, a, b, fTof0, g}, expLocksSize: 9},
-		{span: g0Tog1, expLocks: []roachpb.Span{fTog1Closed, cToEClosed, aToBClosed}, expLocksSize: 9},
-		// Add a key in the middle of a span, which will get merged on commit.
-		{span: c, expLocks: []roachpb.Span{fTog1Closed, cToEClosed, aToBClosed, c}, expLocksSize: 10},
-	}
-	splits := []roachpb.Span{
-		{Key: roachpb.Key("a"), EndKey: roachpb.Key("c")},
-		{Key: roachpb.Key("c"), EndKey: roachpb.Key("f")},
-		{Key: roachpb.Key("f"), EndKey: roachpb.Key("j")},
-	}
-	descs := []roachpb.RangeDescriptor{testMetaRangeDescriptor}
-	for i, s := range splits {
-		descs = append(descs, roachpb.RangeDescriptor{
-			RangeID:          roachpb.RangeID(2 + i),
-			StartKey:         roachpb.RKey(s.Key),
-			EndKey:           roachpb.RKey(s.EndKey),
-			InternalReplicas: []roachpb.ReplicaDescriptor{{NodeID: 1, StoreID: 1}},
-		})
-	}
-	descDB := mockRangeDescriptorDBForDescs(descs...)
-	s := createTestDB(t)
-	st := s.Store.ClusterSettings()
-	trackedWritesMaxSize.Override(ctx, &st.SV, 10) /* 10 bytes and it will condense */
-	defer s.Stop()
-
-	// Check end transaction locks, which should be condensed and split
-	// at range boundaries.
-	expLocks := []roachpb.Span{aToBClosed, cToEClosed, fTog1Closed}
-	sendFn := func(_ context.Context, ba roachpb.BatchRequest) (*roachpb.BatchResponse, error) {
-		resp := ba.CreateReply()
-		resp.Txn = ba.Txn
-		if req, ok := ba.GetArg(roachpb.EndTxn); ok {
-			if !req.(*roachpb.EndTxnRequest).Commit {
-				t.Errorf("expected commit to be true")
-			}
-			et := req.(*roachpb.EndTxnRequest)
-			if a, e := et.LockSpans, expLocks; !reflect.DeepEqual(a, e) {
-				t.Errorf("expected end transaction to have locks %+v; got %+v", e, a)
-			}
-			resp.Txn.Status = roachpb.COMMITTED
-		}
-		return resp, nil
-	}
-	ambient := log.MakeTestingAmbientCtxWithNewTracer()
-	ds := NewDistSender(DistSenderConfig{
-		AmbientCtx: ambient,
-		Clock:      s.Clock,
-		NodeDescs:  s.Gossip,
-		RPCContext: s.Cfg.RPCContext,
-		TestingKnobs: ClientTestingKnobs{
-			TransportFactory: adaptSimpleTransport(sendFn),
-		},
-		RangeDescriptorDB: descDB,
-		Settings:          cluster.MakeTestingClusterSettings(),
-	})
-	tsf := NewTxnCoordSenderFactory(
-		TxnCoordSenderFactoryConfig{
-			AmbientCtx: ambient,
-			Settings:   st,
-			Clock:      s.Clock,
-			Stopper:    s.Stopper(),
-		},
-		ds,
-	)
-	db := kv.NewDB(ambient, tsf, s.Clock, s.Stopper())
-
-	txn := kv.NewTxn(ctx, db, 0 /* gatewayNodeID */)
-	// Disable txn pipelining so that all write spans are immediately
-	// added to the transaction's lock footprint.
-	if err := txn.DisablePipelining(); err != nil {
-		t.Fatal(err)
-	}
-	for i, tc := range testCases {
-		if tc.span.EndKey != nil {
-			if _, err := txn.DelRange(ctx, tc.span.Key, tc.span.EndKey, false /* returnKeys */); err != nil {
-				t.Fatal(err)
-			}
-		} else {
-			if err := txn.Put(ctx, tc.span.Key, []byte("value")); err != nil {
-				t.Fatal(err)
-			}
-		}
-		tcs := txn.Sender().(*TxnCoordSender)
-		locks := tcs.interceptorAlloc.txnPipeliner.lockFootprint.asSlice()
-		if a, e := locks, tc.expLocks; !reflect.DeepEqual(a, e) {
-			t.Errorf("%d: expected keys %+v; got %+v", i, e, a)
-		}
-		locksSize := int64(0)
-		for _, i := range locks {
-			locksSize += int64(len(i.Key) + len(i.EndKey))
-		}
-		if a, e := locksSize, tc.expLocksSize; a != e {
-			t.Errorf("%d: keys size expected %d; got %d", i, e, a)
-		}
-	}
-
-	metrics := txn.Sender().(*TxnCoordSender).Metrics()
-	require.Equal(t, int64(1), metrics.TxnsWithCondensedIntents.Count())
-	require.Equal(t, int64(1), metrics.TxnsWithCondensedIntentsGauge.Value())
-
-	if err := txn.Commit(ctx); err != nil {
-		t.Fatal(err)
-	}
-	require.Zero(t, metrics.TxnsWithCondensedIntentsGauge.Value())
-}
-
 // TestTxnCoordSenderCondenseLockSpans2 verifies that lock spans are condensed
 // along range boundaries when they exceed the maximum intent bytes threshold.
 func TestTxnPipelinerCondenseLockSpans2(t *testing.T) {
@@ -1710,7 +1570,7 @@ func TestTxnPipelinerCondenseLockSpans2(t *testing.T) {
 					EndKey:   roachpb.RKey("d"),
 				}))
 			tp, mockSender := makeMockTxnPipeliner(rangeIter)
-			trackedWritesMaxSize.Override(ctx, &tp.st.SV, tc.maxBytes)
+			TrackedWritesMaxSize.Override(ctx, &tp.st.SV, tc.maxBytes)
 
 			for _, sp := range tc.lockSpans {
 				tp.lockFootprint.insert(roachpb.Span{Key: roachpb.Key(sp.start), EndKey: roachpb.Key(sp.end)})
@@ -1901,7 +1761,7 @@ func TestTxnPipelinerRejectAboveBudget(t *testing.T) {
 			}
 
 			tp, mockSender := makeMockTxnPipeliner(nil /* iter */)
-			trackedWritesMaxSize.Override(ctx, &tp.st.SV, 10) /* reject when exceeding 10 bytes */
+			TrackedWritesMaxSize.Override(ctx, &tp.st.SV, 10) /* reject when exceeding 10 bytes */
 			rejectTxnOverTrackedWritesBudget.Override(ctx, &tp.st.SV, true)
 
 			txn := makeTxnProto()
